updating Movie Entity

+++ src/main/java/com/spring/security/demo/models/Movie.java
@@ -40,7 +40,7 @@ public class Movie {
 
     @Getter
     @Setter
-    private Float rating;
+    private Integer rating;
 
 
     @ManyToMany(fetch = FetchType.LAZY)
